refactor: move hamburger state types into schema file
